This is almost a final version of this project
Everything is in a working state. Things to fix: still norminette errors, need to shorten lines and functions, some files have too many functions in them; there are potential leaks when the game is running, but I am honestly not sure if it's on my end or mlx because I'm freeing everything that was malloc'd; still using illegal strcmp function instead of ft_ one, need to either make a strcmp for my libft or change the code a bit to work with strncmp with I already have; I might still add an end game screen of a sort?